(feat): refactor out grid integration procedure
